+++ native/permissions.js
@@ -1,0 +1,15 @@
+import { RCTModule } from 'react-native-dom';
+
+import ExpoPermissions from '../../expo-permissions/src/ExpoPermissions.web';
+
+export default class Permissions extends RCTModule {
+  static moduleName = 'ExponentPermissions';
+
+  $$getAsync(...types) {
+    return ExpoPermissions.getAsync(...types);
+  }
+
+  $$askAsync(...types) {
+    return ExpoPermissions.askAsync(...types);
+  }
+}